# Swagger/api_mod.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "API",
    "description": "API for Portal System in AEM",
    "termsOfService": "None"
  },
  "host": "otvcrm01:84",
  "basePath": "/",
  "schemes": [ "http", "https" ],
  "consumes": [ "application/json" ],
  "produces": [ "application/json" ],
  "paths": {
...
    "/api/Contacts": {
      "get": {
        "tags": [ "Contacts" ],
        "operationId": "ApiContactsGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Contact" }
            }
          }
        }
      },
      "post": {
        "tags": [ "Contacts" ],
        "operationId": "ApiContactsPost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "contact",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/Contact" }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Contact" }
          }
        }
      }
    },
    "/api/Contacts/getbyemail/{id}/{email}": {
      "get": {
        "tags": [ "Contacts" ],
        "operationId": "ApiContactsGetbyemailByIdByEmailGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Contact" }
          }
        }
      }
    },
    "/api/Contacts/answer/{id}/{answer}": {
      "get": {
        "tags": [ "Contacts" ],
        "operationId": "ApiContactsAnswerByIdByAnswerGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "answer",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Contact" }
          }
        }
      }
    },
    "/api/Contacts/dob/{id}/{dob}": {
      "get": {
        "tags": [ "Contacts" ],
        "operationId": "ApiContactsDobByIdByDobGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "dob",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Contact" }
          }
        }
      }
    },
    "/api/Contacts/{id}": {
      "get": {
        "tags": [ "Contacts" ],
        "operationId": "ApiContactsByIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Contact" }
          }
        }
      },
      "put": {
        "tags": [ "Contacts" ],
        "operationId": "ApiContactsByIdPut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "contact",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/Contact" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      },
      "delete": {
        "tags": [ "Contacts" ],
        "operationId": "ApiContactsByIdDelete",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Contact" }
          }
        }
      }
    },
    "/api/CostCategories": {
      "get": {
        "tags": [ "CostCategories" ],
        "operationId": "ApiCostCategoriesGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/CostCategory" }
            }
          }
        }
      },
      "post": {
        "tags": [ "CostCategories" ],
        "operationId": "ApiCostCategoriesPost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "costCategory",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/CostCategory" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      }
    },
    "/api/CostCategories/{id}": {
      "get": {
        "tags": [ "CostCategories" ],
        "operationId": "ApiCostCategoriesByIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/CostCategory" }
          }
        }
      },
      "put": {
        "tags": [ "CostCategories" ],
        "operationId": "ApiCostCategoriesByIdPut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "costCategory",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/CostCategory" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      },
      "delete": {
        "tags": [ "CostCategories" ],
        "operationId": "ApiCostCategoriesByIdDelete",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/CostCategory" }
          }
        }
      }
    },
    "/api/EligibilityCriterias": {
      "get": {
        "tags": [ "EligibilityCriterias" ],
        "operationId": "ApiEligibilityCriteriasGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/EligibilityCriteria" }
            }
          }
        }
      },
      "post": {
        "tags": [ "EligibilityCriterias" ],
        "operationId": "ApiEligibilityCriteriasPost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "eligibilityCriteria",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/EligibilityCriteria" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      }
    },
    "/api/EligibilityCriterias/{id}": {
      "get": {
        "tags": [ "EligibilityCriterias" ],
        "operationId": "ApiEligibilityCriteriasByIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/EligibilityCriteria" }
          }
        }
      },
      "put": {
        "tags": [ "EligibilityCriterias" ],
        "operationId": "ApiEligibilityCriteriasByIdPut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "eligibilityCriteria",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/EligibilityCriteria" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      },
      "delete": {
        "tags": [ "EligibilityCriterias" ],
        "operationId": "ApiEligibilityCriteriasByIdDelete",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/EligibilityCriteria" }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Account": {
      "type": "object",
      "properties": {
        "accountId": {
          "format": "uuid",
          "type": "string"
        },
        "gcimsClientID": { "type": "string" },
        "accountName": { "type": "string" },
        "legalName": { "type": "string" },
        "primaryAccountAddressId": {
          "format": "uuid",
          "type": "string"
        },
        "paymentAccountAddressId": {
          "format": "uuid",
          "type": "string"
        },
        "accountAddresses": {
          "type": "array",
          "items": { "$ref": "#/definitions/AccountAddress" }
        },
        "accountContacts": {
          "type": "array",
          "items": { "$ref": "#/definitions/AccountContact" }
        },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "internalUpdatedBy": { "$ref": "#/definitions/InternalUser" },
        "createdByUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByUserId": {
          "format": "uuid",
          "type": "string"
        },
        "createdBy": { "$ref": "#/definitions/User" },
        "updatedBy": { "$ref": "#/definitions/User" }
      }
    },
...
    "AccountContact": {
      "type": "object",
      "properties": {
        "accountContactId": {
          "format": "uuid",
          "type": "string"
        },
        "accountId": {
          "format": "uuid",
          "type": "string"
        },
        "account": { "$ref": "#/definitions/Account" },
        "contactId": {
          "format": "uuid",
          "type": "string"
        },
        "contact": { "$ref": "#/definitions/Contact" },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "internalUpdatedBy": { "$ref": "#/definitions/InternalUser" },
        "createdByUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByUserId": {
          "format": "uuid",
          "type": "string"
        },
        "createdBy": { "$ref": "#/definitions/User" },
        "updatedBy": { "$ref": "#/definitions/User" }
      }
    },
    "InternalUser": {
      "type": "object",
      "properties": {
        "internalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "userName": { "type": "string" },
        "firstName": { "type": "string" },
        "lastName": { "type": "string" },
        "email": { "type": "string" },
        "language": { "type": "string" },
        "isPortalAdministrator": { "type": "boolean" },
        "isFundingOpportunityAdministrator": { "type": "boolean" },
        "isWorkloadManager": { "type": "boolean" },
        "isSubmissionReviewer": { "type": "boolean" },
        "fullName": { "type": "string" },
        "roles": { "type": "string" },
        "fundingProgramInternalUsers": {
          "type": "array",
          "items": { "$ref": "#/definitions/FundingProgramInternalUser" }
        },
        "fundingOpportunityInternalUsers": {
          "type": "array",
          "items": { "$ref": "#/definitions/FundingOpportunityInternalUser" }
        },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "userId": {
          "format": "uuid",
          "type": "string"
        },
        "pai": { "type": "string" },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "internalUpdatedBy": { "$ref": "#/definitions/InternalUser" },
        "createdByUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByUserId": {
          "format": "uuid",
          "type": "string"
        },
        "createdBy": { "$ref": "#/definitions/User" },
        "updatedBy": { "$ref": "#/definitions/User" }
      }
    },
    "Address": {
      "type": "object",
      "properties": {
        "addressId": {
          "format": "uuid",
          "type": "string"
        },
        "gcimsAddressId": {
          "format": "int32",
          "type": "integer"
        },
        "addressLine1": { "type": "string" },
        "addressLine2": { "type": "string" },
        "addressLine3": { "type": "string" },
        "addressLine4": { "type": "string" },
        "city": { "type": "string" },
        "gcimsCityID": {
          "format": "int32",
          "type": "integer"
        },
        "province": { "type": "string" },
        "gcimsProvinceID": {
          "format": "int32",
          "type": "integer"
        },
        "country": { "type": "string" },
        "gcimsCountryID": {
          "format": "int32",
          "type": "integer"
        },
        "postal": { "type": "string" },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "internalUpdatedBy": { "$ref": "#/definitions/InternalUser" },
        "createdByUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByUserId": {
          "format": "uuid",
          "type": "string"
        },
        "createdBy": { "$ref": "#/definitions/User" },
        "updatedBy": { "$ref": "#/definitions/User" }
      }
    },
    "Contact": {
      "type": "object",
      "properties": {
        "contactId": {
          "format": "uuid",
          "type": "string"
        },
        "gcimsContactID": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": { "type": "string" },
        "lastName": { "type": "string" },
        "fullName": { "type": "string" },
        "email": { "type": "string" },
        "phoneNumber": { "type": "string" },
        "salutationID": { "type": "string" },
        "preferredLanguageID": { "type": "string" },
        "dateOfBirth": {
          "format": "date-time",
          "type": "string"
        },
        "shareSecretQuestion": { "type": "string" },
        "sharedSecretAnswer": { "type": "string" },
        "isPrimary": { "type": "boolean" },
        "primaryAccountAddressId": {
          "format": "uuid",
          "type": "string"
        },
        "paymentAccountAddressId": {
          "format": "uuid",
          "type": "string"
        },
        "contactAddresses": {
          "type": "array",
          "items": { "$ref": "#/definitions/ContactAddress" }
        },
        "contactAccounts": {
          "type": "array",
          "items": { "$ref": "#/definitions/AccountContact" }
        },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "internalUpdatedBy": { "$ref": "#/definitions/InternalUser" },
        "createdByUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByUserId": {
          "format": "uuid",
          "type": "string"
        },
        "createdBy": { "$ref": "#/definitions/User" },
        "updatedBy": { "$ref": "#/definitions/User" }
      }
    },
...
    "ContactAddress": {
      "type": "object",
      "properties": {
        "contactAddressId": {
          "format": "uuid",
          "type": "string"
        },
        "contact": { "$ref": "#/definitions/Contact" },
        "address": { "$ref": "#/definitions/Address" },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "internalUpdatedBy": { "$ref": "#/definitions/InternalUser" },
        "createdByUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByUserId": {
          "format": "uuid",
          "type": "string"
        },
        "createdBy": { "$ref": "#/definitions/User" },
        "updatedBy": { "$ref": "#/definitions/User" }
      }
    },
...
    "CostCategory": {
      "type": "object",
      "properties": {
        "costCategoryId": {
          "format": "uuid",
          "type": "string"
        },
        "gcimsCostCategoryID": { "type": "string" },
        "titleE": { "type": "string" },
        "titleF": { "type": "string" },
        "lang": { "type": "string" },
        "title": { "type": "string" },
        "toolTip": { "type": "string" },
        "toolTipE": { "type": "string" },
        "toolTipF": { "type": "string" },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "EligibilityCriteria": {
      "type": "object",
      "properties": {
        "eligibilityCriteriaId": {
          "format": "uuid",
          "type": "string"
        },
        "lang": { "type": "string" },
        "description": { "type": "string" },
        "descriptionE": { "type": "string" },
        "descriptionF": { "type": "string" },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "fundingOpportunityEligibilityCriterias": {
          "type": "array",
          "items": { "$ref": "#/definitions/FundingOpportunityEligibilityCriteria" }
        }
      }
    },
    "FundingOpportunityEligibilityCriteria": {
      "type": "object",
      "properties": {
        "fundingOpportunityEligibilityCriteriaId": {
          "format": "uuid",
          "type": "string"
        },
        "fundingOpportunityId": {
          "format": "uuid",
          "type": "string"
        },
        "eligibilityCriteriaId": {
          "format": "uuid",
          "type": "string"
        },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {}
}